requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
cssselect==1.1.0
cycler==0.10.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.2.0
idna==3.2
jieba3k==0.35.1
joblib==1.0.1
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.4.3
newspaper3k==0.2.8
nltk==3.6.3
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
pip==21.2.4
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests-file==1.5.1
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
setuptools==57.0.0
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
snscrape==0.3.4
soupsieve==2.2.1
stopwords==1.0.0
threadpoolctl==2.2.0
tinysegmenter==0.3
tldextract==3.1.2
tqdm==4.62.2
urllib3==1.26.6
wheel==0.36.2
